Add some more useful combinators
- A predicate combinator, which returns true if a character matches a
  passed function.
- Zero or more/One or more of another parser.
- Left/Right, which discard the opposite corresponding value of a pair,
  useful for flattening the result of and_then.

There may be redundant parsers in this list, e.g possibly until, which
might be replaceable with a combination of the zero/one or more +
a new not combinator + predicate, but this should be good for now to
implement some complete parsers.

Need to figure out a better system of error messages.